requirements.txt
# FetalAgent core dependencies
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
autogen-core>=0.4.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
pandas>=1.3.0

# External tool dependencies (may require separate conda environments)
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.6.0
timm>=0.6.0
transformers>=4.20.0
SimpleITK>=2.1.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.5.0
open_clip_torch>=2.0.0
einops>=0.4.0
monai>=1.0.0

# nnU-Net (for HC and stomach segmentation tools)
# Install separately: pip install nnunetv2
